AbstractIdentifiableDescriptor: include builder
