+++ model/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/model/descriptor/impl/AbstractIdentifiableDescriptor.java
@@ -18,6 +18,7 @@ import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.AdministrationDesc
 import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.DescriptionDescriptor;
 import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.EndpointDescriptor;
 import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.IdentificationDescriptor;
+import io.adminshell.aas.v3.model.builder.ExtendableBuilder;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -34,7 +35,7 @@ public abstract class AbstractIdentifiableDescriptor {
     private List<DescriptionDescriptor> descriptions;
     private IdentificationDescriptor identification;
 
-    public AbstractIdentifiableDescriptor() {
+    protected AbstractIdentifiableDescriptor() {
         idShort = null;
         endpoints = new ArrayList<>();
         administration = null;
@@ -115,4 +116,47 @@ public abstract class AbstractIdentifiableDescriptor {
         return Objects.hash(idShort, endpoints, administration, descriptions, identification);
     }
 
+    public abstract static class AbstractBuilder<T extends AbstractIdentifiableDescriptor, B extends AbstractBuilder<T, B>> extends ExtendableBuilder<T, B> {
+
+        public B idShort(String value) {
+            getBuildingInstance().setIdShort(value);
+            return getSelf();
+        }
+
+
+        public B endpoints(List<EndpointDescriptor> value) {
+            getBuildingInstance().setEndpoints(value);
+            return getSelf();
+        }
+
+
+        public B endpoint(EndpointDescriptor value) {
+            getBuildingInstance().getEndpoints().add(value);
+            return getSelf();
+        }
+
+
+        public B administration(AdministrationDescriptor value) {
+            getBuildingInstance().setAdministration(value);
+            return getSelf();
+        }
+
+
+        public B descriptions(List<DescriptionDescriptor> value) {
+            getBuildingInstance().setDescriptions(value);
+            return getSelf();
+        }
+
+
+        public B oneDescription(DescriptionDescriptor value) {
+            getBuildingInstance().getDescriptions().add(value);
+            return getSelf();
+        }
+
+
+        public B identification(IdentificationDescriptor value) {
+            getBuildingInstance().setIdentification(value);
+            return getSelf();
+        }
+    }
 }
