+++ model/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/model/descriptor/impl/DefaultAssetAdministrationShellDescriptor.java
@@ -16,8 +16,6 @@ package de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.impl;
 
 import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.AdministrationDescriptor;
 import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.AssetAdministrationShellDescriptor;
-import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.DescriptionDescriptor;
-import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.EndpointDescriptor;
 import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.IdentificationDescriptor;
 import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.IdentifierKeyValuePairDescriptor;
 import de.fraunhofer.iosb.ilt.faaast.service.model.descriptor.ReferenceDescriptor;
@@ -28,7 +26,6 @@ import io.adminshell.aas.v3.model.KeyElements;
 import io.adminshell.aas.v3.model.KeyType;
 import io.adminshell.aas.v3.model.Reference;
 import io.adminshell.aas.v3.model.Submodel;
-import io.adminshell.aas.v3.model.builder.ExtendableBuilder;
 import io.adminshell.aas.v3.model.impl.DefaultKey;
 import io.adminshell.aas.v3.model.impl.DefaultReference;
 import java.util.ArrayList;
@@ -116,52 +113,11 @@ public class DefaultAssetAdministrationShellDescriptor extends AbstractIdentifia
         return new Builder();
     }
 
-    public abstract static class AbstractBuilder<T extends DefaultAssetAdministrationShellDescriptor, B extends AbstractBuilder<T, B>> extends ExtendableBuilder<T, B> {
-
-        public B idShort(String value) {
-            getBuildingInstance().setIdShort(value);
-            return getSelf();
-        }
-
-
-        public B endpoints(List<EndpointDescriptor> value) {
-            getBuildingInstance().setEndpoints(value);
-            return getSelf();
-        }
-
-
-        public B endpoint(EndpointDescriptor value) {
-            getBuildingInstance().getEndpoints().add(value);
-            return getSelf();
-        }
-
-
-        public B administration(AdministrationDescriptor value) {
-            getBuildingInstance().setAdministration(value);
-            return getSelf();
-        }
-
-
-        public B descriptions(List<DescriptionDescriptor> value) {
-            getBuildingInstance().setDescriptions(value);
-            return getSelf();
-        }
-
-
-        public B oneDescription(DescriptionDescriptor value) {
-            getBuildingInstance().getDescriptions().add(value);
-            return getSelf();
-        }
-
+    public abstract static class AbstractBuilder<T extends DefaultAssetAdministrationShellDescriptor, B extends AbstractBuilder<T, B>>
+            extends AbstractIdentifiableDescriptor.AbstractBuilder<T, B> {
 
         public B globalAssetId(ReferenceDescriptor value) {
             getBuildingInstance().setGlobalAssetId(value);
-            return getSelf();
-        }
-
-
-        public B identification(IdentificationDescriptor value) {
-            getBuildingInstance().setIdentification(value);
             return getSelf();
         }
